tutorials/deploy-layer-models/streamlit/requirements.txt
numpy
streamlit
tensorflow-cpu==2.6
pillow